requirements.txt
# --- Web Framework & Core ---
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1

# --- Security & Tokens ---
itsdangerous==2.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1

# --- Image Processing & AI ---
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.26.2

# --- Database & Server ---
SQLAlchemy==2.0.23
# Use gunicorn if you plan to deploy to Linux/Production
gunicorn==21.2.0